+++ src/panel/index.js
@@ -1,6 +1,6 @@
 import { createRoot } from "react-dom/client";
 import App from "./App";
-import "./styles.css";
+import "./styles/index.css";
 
 const root = createRoot(document.getElementById("root"));
 root.render(<App />);
