+++ src/panel/styles/base.css
@@ -1,0 +1,3 @@
+/* ──────────────────────────────────────────────────────────────────────────────
+   base.css — reset, design tokens, typography defaults
+   ──────────────────────────────────────────────────────────────────────────── */
