+++ src/panel/styles/checklist.css
@@ -1,0 +1,16 @@
+/* ──────────────────────────────────────────────────────────────────────────────
+   checklist.css — manual testing checklist panel
+   Classes:
+     .checklist-panel, .checklist-header
+     .checklist-progress-row, .checklist-progress-text
+     .checklist-pct, .checklist-pct--{complete|partial|low}
+     .checklist-bar-track, .checklist-bar-fill, .checklist-bar-fill--{complete|partial|low}
+     .checklist-note, .checklist-categories
+     .checklist-category, .checklist-category--done
+     .checklist-cat-header, .checklist-cat-icon, .checklist-cat-name
+     .checklist-cat-count, .checklist-cat-count--done
+     .principle-chevron
+     .checklist-items, .checklist-item, .checklist-item--checked
+     .checklist-checkbox, .checklist-item-body, .checklist-item-text, .checklist-item-wcag
+     .checklist-reset
+   ──────────────────────────────────────────────────────────────────────────── */
