+++ src/panel/styles/contrast.css
@@ -1,0 +1,35 @@
+/* ──────────────────────────────────────────────────────────────────────────────
+   contrast.css — Contrast panel (Page scan / Pick element / Check colours)
+
+   .contrast-panel
+   .mode-toggle, .mode-btn, .mode-btn--active
+
+   Page scan results:
+     .scan-mode, .contrast-scan-results
+     .cscan-hero, .cscan-hero-stats
+     .cscan-stat-card, .cscan-stat-card--fail|warn|pass, .cscan-stat-card--active
+     .cscan-stat-num, .cscan-stat-text
+     .cscan-hero-actions, .cscan-howto-toggle, .cscan-explainer
+     .cscan-group, .cscan-group-header, .cscan-group-icon, .cscan-group-label, .cscan-group-counts
+     .cscan-fail-badge, .cscan-warn-badge, .cscan-total-badge
+     .principle-chevron-icon
+     .cscan-items, .cscan-item, .cscan-item--fail|warn|pass
+     .cscan-preview-row, .cscan-colour-preview, .cscan-preview-text
+     .cscan-preview-meta, .cscan-preview-top
+     .cscan-ratio, .cscan-ratio--fail|warn|pass
+     .cscan-status-pill, .cscan-status-pill--fail|warn|pass
+     .cscan-hex-row, .cscan-swatch-mini, .cscan-hex, .cscan-hex-on, .cscan-size-note
+     .cscan-actions, .cscan-jump-btn, .cscan-verify-btn
+
+   Pick element mode:
+     .picker-mode, .picker-prompt, .picker-waiting, .crosshair-icon
+     .picker-result, .swatch-row, .swatch-item, .swatch, .swatch-label, .swatch-hex
+
+   Check colours mode (manual):
+     .manual-mode, .manual-intro
+     .colour-inputs, .colour-field, .colour-input-row
+     .colour-picker-input, .hex-input, .eyedropper-btn
+     .preview, .preview-normal, .preview-large
+     .ratio-display, .ratio-number, .ratio-number--pass|fail
+     .ratio-grid, .ratio-row, .ratio-label, .ratio-req
+   ──────────────────────────────────────────────────────────────────────────── */
