+++ src/panel/styles/onboarding.css
@@ -1,0 +1,9 @@
+/* ──────────────────────────────────────────────────────────────────────────────
+   onboarding.css — first-run onboarding overlay
+   Classes:
+     .onboarding-overlay, .onboarding-card
+     .onboarding-logo, .onboarding-brand
+     .onboarding-icon, .onboarding-title, .onboarding-body
+     .onboarding-dots, .onboarding-dot, .onboarding-dot--active
+     .onboarding-actions, .onboarding-back, .onboarding-next, .onboarding-skip
+   ──────────────────────────────────────────────────────────────────────────── */
