+++ src/panel/styles/tools.css
@@ -1,0 +1,47 @@
+/* ──────────────────────────────────────────────────────────────────────────────
+   tools.css — Tools panel (Focus + Visuals + Dynamic subtabs)
+
+   .tools-panel
+
+   Focus tab home:
+     .focus-tab-home, .focus-tab-intro, .focus-tool-grid
+     .focus-tool-card, .focus-tool-icon, .focus-tool-title, .focus-tool-desc
+     .info-callout
+
+   Live keyboard test:
+     .focus-mode-panel, .focus-mode-header, .focus-mode-title
+     .focus-mode-instructions, .focus-mode-step, .focus-step-num
+     .focus-current-card, .focus-current-card--ok, --warn, --pending
+     .focus-current-top, .focus-stop-num, .focus-stop-num--ok|warn|pending
+     .focus-stop-tag, .focus-status-row, .focus-status-row--ok|warn|pending
+     .focus-status-note, .focus-current-empty, .focus-empty-icon
+     .focus-stops-list, .focus-stops-list-title, .focus-stop-row
+     .focus-stop-icon, .focus-stop-icon--ok|warn|pending, .focus-stop-n, .focus-stop-el
+
+   Tab order map:
+     .tab-order-panel, .tab-order-header, .tab-order-header-left
+     .tab-order-title, .tab-order-count
+     .tab-order-info-toggle, .tab-order-info-panel
+     .tab-order-issues-bar, .tab-order-ok-bar
+     .tab-order-group, .tab-order-group-header, .tab-order-group-header--issues|ok
+     .tab-order-group-title, .tab-order-group-count, .tab-order-group-count--bad|ok
+     .tab-order-rows
+     .tab-stop-card, .tab-stop-card--ok|warn|danger, .tab-stop-card--selected
+     .tab-stop-card-header, .tab-stop-num, .tab-stop-num--ok|warn|danger
+     .tab-stop-card-info, .tab-stop-card-label
+     .tab-stop-card-type, .tab-stop-card-type--issue
+     .tab-stop-card-actions
+     .tab-stop-detail, .tab-stop-detail-body
+     .tab-stop-ok-detail, .tab-stop-ok-section, .tab-stop-ok-label, .tab-stop-ok-text
+
+   Visuals:
+     .visuals-tab, .visuals-section, .visuals-section-header, .visuals-section-icon
+     .visuals-section-title, .visuals-section-desc
+     .visuals-running, .visuals-result, .visuals-result--ok
+     .visuals-result-icon, .visuals-result-desc
+     .hc-checklist, .hc-checklist-title, .hc-checklist-hint
+     .hc-item, .hc-item-check-btns, .hc-btn, .hc-item-body, .hc-item-label, .hc-note-input
+
+   Dynamic:
+     .dynamic-tab, .dynamic-summary
+   ──────────────────────────────────────────────────────────────────────────── */
